Shadow Emoji rework
Shadow emoji now grants a dash that can go downwards.

Dashing has i-frames but if you dash into an enemy the i-frames disappear for 10 seconds.

Dashing into an enemy with i-frames inflicts them with a debuff that removes all of their defense for 2 seconds

debuff sprites will be done later