Two new debuffs added
Dimensional Shock is a debuff that is now afflicted by the Dimensional God's projectile as well as the Data Miner. More items in the future will receive this debuff.

Accursed is a debuff that is inflicted by the Hex Rapier. Inflicted enemies spawn Accursed Stalkers that home into the nearest enemy.