Monarch expert item rework
Instead of continuesly draining health with a 10% damage increase it now has the effect below.

Potion effectiveness is halved, during potion sickness, life regeneration is disabled and all damage done to the player is collected in an internal value. When Potion Sickness runs out, the player gets insta healed by a quarter of the damage that was collected and the collected damage resets.