Awakened Archbird rework
Is now named Rahlem's Flamescar.

You can right click to cut yourself and deal 20 damage. If you do so, the weapon starts inflicting Blazing Suffering and shoots 2 types of embers.

Green embers pierce but are weaker, Red embers inflict Blazing Suffering but do not pierce.

When the effect ends, the player is inflicted with a cooldown, if they use the right click cut during the cooldown, the demanded health is doubled and the player is inflicted with Blazing Suffering for a short time.