undergrouns chaos biome adjustment
Added two new wall types that look identical to the volcano and lake depths walls. The normal walls now initiate the underground biome when stepped in front of. This was done to prevent the surface biome from overriding the depths and volcano when they generate too far up on the map.